hide the actual stream pool
